requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
requests>=2.31.0
oauthlib>=3.2.2
Flask-Session>=0.5.0
python-jose>=3.3.0
PyJWT>=2.8.0